requirements.txt
# Core Dependencies
python-dotenv==1.0.0
pyyaml==6.0.1
cryptography==41.0.7

# Data & Analysis
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4

# Technical Analysis
TA-Lib==0.4.28
pandas-ta==0.3.14b0

# Broker Integration
kiteconnect==4.3.0

# Database
sqlalchemy==2.0.23
alembic==1.13.1

# Web Framework
fastapi==0.108.0
uvicorn[standard]==0.25.0
python-multipart==0.0.6
websockets==12.0

# Visualization
plotly==5.18.0
matplotlib==3.8.2
seaborn==0.13.0

# Utilities
requests==2.31.0
aiohttp==3.9.1
python-telegram-bot==20.7
schedule==1.2.0
APScheduler==3.10.4

# Testing
pytest==7.4.3
pytest-cov==4.1.0
pytest-asyncio==0.21.1

# Development
black==23.12.1
flake8==7.0.0
mypy==1.7.1
